pyproject.toml
[project]
name = "review-crew"
version = "0.1.0"
description = "Multi-agent content review system using Strands Agents"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "strands-agents>=0.1.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "openai>=1.0.0",  # For LM Studio OpenAI-compatible API support
]
requires-python = ">=3.10"

readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
lm-studio = [
    "openai>=1.0.0",  # For LM Studio support
]
ollama = [
    "ollama>=0.1.0",  # For Ollama support (if available)
]

[project.scripts]
review-crew = "src.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
